Got upload file button working on frontend and refactored
