+++ src/backend/app.py
@@ -1,10 +1,10 @@
 from flask import Flask
 from flask_cors import CORS
 
-from backend.chatbot.text_transformer.neo4j_interactor import Neo4JInteractor
-from backend.chatbot.text_transformer.text_vectoriser import TextVectoriser
-from backend.mongodb.DocumentStore import DocumentStore
-from backend.upload.DocumentUploader import DocumentUploader
+from chatbot.text_transformer.neo4j_interactor import Neo4JInteractor
+from chatbot.text_transformer.text_vectoriser import TextVectoriser
+from mongodb.DocumentStore import DocumentStore
+from upload.DocumentUploader import DocumentUploader
 
 
 def initialise_collection() -> DocumentStore.Collection:
@@ -18,8 +18,8 @@ def initialise_vector_database() -> tuple[Neo4JInteractor, TextVectoriser]:
 
 def register_upload_routes(app: Flask) -> None:
     collection = initialise_collection()
-    vector_db, vectoriser = initialise_vector_database()
-    document_uploader = DocumentUploader(collection, vector_db, vectoriser)
+    #vector_db, vectoriser = initialise_vector_database()
+    document_uploader = DocumentUploader(collection)#, vector_db, vectoriser)
     document_uploader.register_routes(app)
 
 def start_app() -> None:
