+++ src/backend/chatbot/chat_service.py
@@ -1,6 +1,6 @@
 from flask import Flask
 from flask import request, jsonify 
-from backend.chatbot.chat import Chatbot
+from chatbot.chat import Chatbot
 from flask_cors import CORS
 
 app = Flask(__name__)
