+++ src/frontend/file-browser/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import UploadFileButton from './UploadFileButton';
+import UploadFileButton from '../chatbot/src/components/UploadFileButton';
 import BrowseFolderButton from './BrowseFolderButton';
 
 const App: React.FC = () => {
@@ -7,18 +7,10 @@ const App: React.FC = () => {
     console.log("Selected file:", file.name);
     // You can send the file to a backend here
   };
-  const handleFilesSelected = (files: FileList) => {
-    console.log("Selected folder files:");
-    for (const file of files) {
-      console.log("Selected file:", file.name);
-      // You can send each file to a backend here
-    }
-  };
 
   return (
     <div className='App'>
       <UploadFileButton onFileSelected={handleFileSelected} />
-      <BrowseFolderButton onFilesSelected={handleFilesSelected} />
     </div>
   );
 };
